Infer base type better when generating a type
