+++ src/EditorFeatures/CSharpTest/Diagnostics/GenerateType/GenerateTypeTests.cs
@@ -7,9 +7,7 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CodeStyle;
-using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.CodeFixes.GenerateType;
-using Microsoft.CodeAnalysis.CSharp.Diagnostics;
 using Microsoft.CodeAnalysis.CSharp.Test.Utilities;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Test.Utilities;
@@ -5355,32 +5353,50 @@ internal class Class
 }}",
     index: 1);
         }
-    }
-
-    public partial class GenerateTypeWithUnboundAnalyzerTests : AbstractCSharpDiagnosticProviderBasedUserDiagnosticTest
-    {
-        internal override (DiagnosticAnalyzer, CodeFixProvider) CreateDiagnosticProviderAndFixer(Workspace workspace)
-            => (new CSharpUnboundIdentifiersDiagnosticAnalyzer(), new GenerateTypeCodeFixProvider());
-
-        protected override ImmutableArray<CodeAction> MassageActions(ImmutableArray<CodeAction> codeActions)
-            => FlattenActions(codeActions);
 
         [Fact, Trait(Traits.Feature, Traits.Features.CodeActionsGenerateType)]
-        [WorkItem(13211, "https://github.com/dotnet/roslyn/issues/13211")]
-        public async Task TestGenerateOffOfIncompleteMember()
+        [WorkItem(270, "https://github.com/dotnet/roslyn/issues/270")]
+        public async Task TestGenerateInIsExpression()
         {
             await TestInRegularAndScriptAsync(
-@"class Class
+@"using System;
+ 
+class Program
 {
-    public [|Goo|]
+    static void Main(Exception p)
+    {
+        bool result = p is [|SampleType|];
+    }
 }",
-@"class Class
+@"using System;
+using System.Runtime.Serialization;
+
+class Program
 {
-    public Goo
+    static void Main(Exception p)
+    {
+        bool result = p is SampleType;
+    }
 }
 
-internal class Goo
+[Serializable]
+internal class SampleType : Exception
 {
+    public SampleType()
+    {
+    }
+
+    public SampleType(string message) : base(message)
+    {
+    }
+
+    public SampleType(string message, Exception innerException) : base(message, innerException)
+    {
+    }
+
+    protected SampleType(SerializationInfo info, StreamingContext context) : base(info, context)
+    {
+    }
 }",
 index: 1);
         }
